requirements.txt
crewai
crewai-tools
python-dotenv
langchain-google-genai
mistralai
vnstock==3.2.6
vnai==2.1.9
pandas
numpy
streamlit
streamlit-chat
beautifulsoup4
crewai[google-genai]
google
google-generativeai
requests
tenacity
certifi
Flask
Flask-SocketIO
Flask-Cors
eventlet
schedule
weasyprint
markdown2
matplotlib
mplfinance
pdfkit
certifi